Move MixingMatrix into loaders
Move all MixingMatrix and associated classes into loaders, and do
similar things with the tests
